feat(components): add NewIconButton component
