+++ packages/components/src/index.ts
@@ -24,6 +24,10 @@ export * from './components/buttons/ButtonGroup/ButtonGroup';
 export * from './components/buttons/NewButtonGroup/NewButtonGroup';
 export * from './components/buttons/IconButton/IconButton';
 export { NewButton, type NewButtonProps } from './components/buttons/NewButton/NewButton';
+export {
+    NewIconButton,
+    type NewIconButtonProps,
+} from './components/buttons/NewIconButton/NewIconButton';
 export * from './components/Icon/Icon';
 export { ComponentWithSubIcon } from './components/ComponentWithSubIcon/ComponentWithSubIcon';
 export * from './components/buttons/PinButton/PinButton';
